Fix crash when tapping on jump list letters
The app crashes when user taps on the jump list letters on the series pivot.
This is due to the change from handling selection change to handling tap
events. When user taps on jump list, the code thinks that the user wants to
view a particular series even though that's not the case.

The fix is to not navigate to series' chapters page if the selected item is
null.